reduce memory of sequence from int to byte

+++ src/headers/utils.h
@@ -9,14 +9,14 @@ template<typename T>
 T* oalloc(int count){
     return (T*)malloc(count*sizeof(T));
 }
-int mapping_Char(char c){
-    for(int i=0;i<CHAR_NUMBER;i++){
+byte mapping_Char(char c){
+    for(byte i=0;i<CHAR_NUMBER;i++){
         if(c==protected_space::Char_map[i])return i;
     }
     printf("Char [%c] not found!!\n",c);
     exit(0);
 }
-inline char to_Char(int i){
+inline char to_Char(byte i){
     return protected_space::Char_map[i];
 }
 #endif
